Fix hashes collision in generated Pipfile.locks
When multiple stacks were generated, there was an issue with same hashes being
used multiple times for different packages. This patch addresses the this issue
- when python solver implemented on top of graph database did resolution,
constructor of PackageVersion class was used with arguments from the original
package. This caused reference to the same list being passed to different
package of same name - the gathering of package hashes did checks for same list
reference.