Implement structure for dependency graph adjustment
Removing a package from dependency graph took too much time when done on
large stacks, such as TensorFlow. The issue is in checking all the paths (20M
in case of TensorFlow stacks). This implementation constructs in-memory graph
for representing how dependency graph looks like with efficient transactional
package removal and package scoring - only affected sub-graph is checked not to
iterate on 20M records in case of TensorFlow.