Introduce heat up phase for the latest predictor
This phase will generate new states out of the initial state. The newly
generated states are always generated from the latest possible direct
dependnecy found in the stack. This way we minimalize a possibility to of
getting stuck during predictions in a local area in the state space where we do
not lead to any final state. This addresses the OOM issue spotted in the stage
environment.